cms/requirements.txt
# FastAPI 核心依赖
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# 文件处理
PyYAML==6.0.1
python-multipart==0.0.6
Pillow==10.2.0

# Git 集成
GitPython==3.1.41

# 工具库
python-dateutil==2.8.2
aiofiles==23.2.1

# Google Scholar 抓取（可选）
scholarly==1.7.11
beautifulsoup4==4.12.3
requests==2.31.0

# 文件监控（Overleaf PDF 监控）
watchdog==4.0.0

# 调度任务
APScheduler==3.10.4
